utils/docai-wh/requirements.txt
google-cloud-logging
google-cloud-documentai==2.16.1
google-cloud-storage
google-cloud-contentwarehouse==0.5.1